Document reloadable configs (#125)
Closes https://github.com/DefinedNet/nebula-docs/issues/19

We didn't have any good listing of which options can be reloaded via a
SIGHUP vs those that require a full restart to take effect. This adds a
simple badge to indicate which options can be reloaded. AFAICT,
`tun.mtu` is only reloadable on linux, but that's the only difference
between platforms that I could find.